+++ src/main/java/com/yooksi/betterarchery/common/BetterArchery.java
@@ -1,26 +1,39 @@
 package com.yooksi.betterarchery.common;
 
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-
+import net.minecraft.creativetab.CreativeTabs;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
-@Mod(modid = BetterArcheryReborn .MODID, version = BetterArcheryReborn.VERSION, name= BetterArcheryReborn.NAME, acceptedMinecraftVersions = "[1.10,1.10.2]")
+@Mod(modid = BetterArchery .MODID, version = BetterArchery.VERSION, name= BetterArchery.NAME, acceptedMinecraftVersions = "[1.10,1.10.2]")
 
-public class BetterArcheryReborn 
+public class BetterArchery 
 {
 	public static final String MODID = "betterarchery";
 	public static final String NAME = "Better Archery Reborn";
-    public static final String VERSION = "0.1.0";
-    
-    /** The instance of our mod that Forge uses. */
-	@Mod.Instance(MODID)
-	public static BetterArcheryReborn  instance;
+    public static final String VERSION = "0.1.1";
 
+	// This is where all our custom items should be listed in-game
+	public static final CreativeTabs tabBetterArchery = new CreativeTabs("Better Archery") 
+	{
+		@Override
+		@SideOnly(Side.CLIENT)
+	    public net.minecraft.item.Item getTabIconItem()
+		{ 
+			return net.minecraft.init.Items.BOW; 
+		}
+	};
+	
+	@Mod.Instance(MODID)
+	public static BetterArchery  instance;
+	
+	@SidedProxy(clientSide = "com.yooksi.betterarchery.network.ClientProxy",   
+			serverSide = "com.yooksi.betterarchery.network.ServerProxy")
+	
     public static CommonProxy proxy = new CommonProxy();
 	
 	/**