+++ src/main/java/com/yooksi/betterarchery/network/ClientProxy.java
@@ -1,0 +1,17 @@
+package com.yooksi.betterarchery.network;
+
+import com.yooksi.betterarchery.common.CommonProxy;
+
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
+
+public class ClientProxy extends CommonProxy
+{
+	@Override
+	@SideOnly(Side.CLIENT)
+	public void preInit(FMLPreInitializationEvent event)
+	{
+		super.preInit(event);
+	}
+}