+++ src/main/java/com/yooksi/betterarchery/network/ServerProxy.java
@@ -1,0 +1,5 @@
+package com.yooksi.betterarchery.network;
+
+public class ServerProxy {
+
+}