+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this mod project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and uses a custom versioning system.
 
+## Version 0.1.1 (2016-10-09)
+### Added
+- Network support for client and server.
+- Custom mod creative tab.
+
+### Changed
+- Main mod class name.
+
 ## Version 0.1.0 (2016-10-09)
 ### Added
 - A readme file.